recomputing fingerprints after (hopefully) fixing quail "bug"
note: the "bug" was an edge case related to computing permutation-corrected clustering scores.  in the previous implementation, if all list items had the same feature value for a given feature, the corrected clustering score for that feature would be 1, since the comparisons for permuted vs. observed scores were made using a strictly "less than" comparison (and none of the permuted scores would be greater than the observed score).

to fix this edge case, quail now considers 3 cases for each permuted score (and computed the average across permutations to get the corrected score):
- case 1: permuted score < observed score -- this is worth 1 point
- case 2: permuted score > observed score -- this is worth 0 points
- case 3: permuted score == observed score -- this is *now* worth 0.5 points

now the "all features have the same value" scenario should result in a corrected score of 0.5, which makes more "sense."